Fixed liquibase settings

+++ src/main/java/com/example/bookshopapp/data/tag/TagEntity.java
@@ -14,9 +14,10 @@ import java.util.Set;
 @Getter
 @Setter
 public class TagEntity {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     private String name;
